Update patch set 1
Patch Set 1:

I can consistently make the browser FC with this in. I'm still testing if it's one of my other patches to my own build but it may need a more thorough test before it is included but I haven't encountered an FC with the browser before this was put in so I'm inclined to believe this caused it.

*got the libwebkitaccel.so from the Ace nightly if that's important

Patch-set: 1