Update patch set 1
Patch Set 1:

Revert for the other devices you mean? I see no point in removing support for turning it on with the flag, like this commit does so that those of us who don't care about breaking a few apps can easily apply it in our own personal builds.

Will be interesting to see how HTC and the others deal with it.

Patch-set: 1